+++ .gitignore
@@ -1,0 +1,3 @@
+src/*.o
+bin/zumo.*
+*.swp
