--- src/.gitignore
@@ -1,3 +1,0 @@
-main.o
-zumo.*
-Makefile
