+++ src/headers/usart.h
@@ -1,0 +1,8 @@
+#ifndef USART_H
+#define USART_H
+
+void USART_init();
+void writeChar(char data);
+void writeString(char* Pstring);
+
+#endif
